// src/client/Pg.DataverseSync.Admin/src/generated/models/EnvironmentvariablevaluesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Environmentvariablevaluescomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'Deleted Unpublished'
} as const;
export type Environmentvariablevaluescomponentstate = keyof typeof Environmentvariablevaluescomponentstate;
export const Environmentvariablevaluesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Environmentvariablevaluesstatecode = keyof typeof Environmentvariablevaluesstatecode;
export const Environmentvariablevaluesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Environmentvariablevaluesstatuscode = keyof typeof Environmentvariablevaluesstatuscode;

export interface EnvironmentvariablevaluesBase {
  "EnvironmentVariableDefinitionId@odata.bind": string;
  environmentvariablevalueid: string;
  importsequencenumber?: number;
  introducedversion?: string;
  iscustomizable: string;
  overriddencreatedon?: string;
  schemaname: string;
  statecode: Environmentvariablevaluesstatecode;
  statuscode?: Environmentvariablevaluesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  value?: string;
}

export interface Environmentvariablevalues extends EnvironmentvariablevaluesBase {
  componentstate: Environmentvariablevaluescomponentstate;
  componentstatename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  environmentvariabledefinitionidname?: string;
  environmentvariablevalueidunique: string;
  ismanaged: boolean;
  ismanagedname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overwritetime: string;
  ownerid: string;
  owneridtype: string;
  solutionid: string;
  statecodename?: string;
  statuscodename?: string;
  supportingsolutionid?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  environmentvariabledefinitionid?: object;
  _environmentvariabledefinitionid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}