// src/client/Pg.DataverseSync.Admin/src/generated/models/ServiceendpointsModel.ts
export const Serviceendpointsauthtype = {
  0: 'Not Specified',
  8: 'Access Key',
  9: 'Managed Identity',
  1: 'ACS',
  2: 'SAS Key',
  3: 'SAS Token',
  4: 'Webhook Key',
  5: 'Http Header',
  6: 'Http Query String',
  7: 'Connection String'
} as const;
export type Serviceendpointsauthtype = keyof typeof Serviceendpointsauthtype;
export const Serviceendpointscomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'Deleted Unpublished'
} as const;
export type Serviceendpointscomponentstate = keyof typeof Serviceendpointscomponentstate;
export const Serviceendpointsconnectionmode = {
  1: 'Normal',
  2: 'Federated'
} as const;
export type Serviceendpointsconnectionmode = keyof typeof Serviceendpointsconnectionmode;
export const Serviceendpointscontract = {
  1: 'OneWay',
  2: 'Queue',
  3: 'Rest',
  4: 'TwoWay',
  5: 'Topic',
  6: 'Queue (Persistent)',
  7: 'Event Hub',
  8: 'Webhook',
  9: 'Event Grid',
  10: 'Managed Data Lake',
  11: 'Container Storage'
} as const;
export type Serviceendpointscontract = keyof typeof Serviceendpointscontract;
export const Serviceendpointsmessagecharset = {
  0: 'Default',
  1: 'UTF8',
  2: 'Windows1252'
} as const;
export type Serviceendpointsmessagecharset = keyof typeof Serviceendpointsmessagecharset;
export const Serviceendpointsmessageformat = {
  1: 'Binary XML',
  2: 'Json',
  3: 'Text XML'
} as const;
export type Serviceendpointsmessageformat = keyof typeof Serviceendpointsmessageformat;
export const Serviceendpointsnamespaceformat = {
  1: 'Namespace Name',
  2: 'Namespace Address'
} as const;
export type Serviceendpointsnamespaceformat = keyof typeof Serviceendpointsnamespaceformat;
export const Serviceendpointsschematype = {
  1: 'Event Grid',
  2: 'Cloud Events'
} as const;
export type Serviceendpointsschematype = keyof typeof Serviceendpointsschematype;
export const Serviceendpointsuserclaim = {
  1: 'None',
  2: 'UserId',
  3: 'UserInfo'
} as const;
export type Serviceendpointsuserclaim = keyof typeof Serviceendpointsuserclaim;

export interface ServiceendpointsBase {
  authtype?: Serviceendpointsauthtype;
  authvalue?: string;
  connectionmode: Serviceendpointsconnectionmode;
  contract: Serviceendpointscontract;
  description?: string;
  introducedversion?: string;
  iscustomizable: string;
  "KeyVaultReferenceId@odata.bind"?: string;
  "ManagedIdentityId@odata.bind"?: string;
  messagecharset?: Serviceendpointsmessagecharset;
  messageformat?: Serviceendpointsmessageformat;
  name: string;
  namespaceaddress?: string;
  namespaceformat?: Serviceendpointsnamespaceformat;
  path: string;
  runtimeintegrationproperties?: string;
  saskey?: string;
  saskeyname?: string;
  sastoken?: string;
  schematype?: Serviceendpointsschematype;
  serviceendpointid: string;
  solutionnamespace: string;
  url?: string;
  usekeyvaultconfiguration?: boolean;
  userclaim: Serviceendpointsuserclaim;
}